+++ Tests/TuistKitTests/Commands/ScaffoldCommandTests.swift
@@ -17,15 +17,18 @@ final class ScaffoldCommandTests: TuistUnitTestCase {
     var parser: ArgumentParser!
     var templateLoader: MockTemplateLoader!
     var templatesDirectoryLocator: MockTemplatesDirectoryLocator!
+    var templateGenerator: MockTemplateGenerator!
 
     override func setUp() {
         super.setUp()
         parser = ArgumentParser.test()
         templateLoader = MockTemplateLoader()
         templatesDirectoryLocator = MockTemplatesDirectoryLocator()
+        templateGenerator = MockTemplateGenerator()
         subject = ScaffoldCommand(parser: parser,
                                   templateLoader: templateLoader,
-                                  templatesDirectoryLocator: templatesDirectoryLocator)
+                                  templatesDirectoryLocator: templatesDirectoryLocator,
+                                  templateGenerator: templateGenerator)
     }
 
     override func tearDown() {
@@ -33,6 +36,7 @@ final class ScaffoldCommandTests: TuistUnitTestCase {
         subject = nil
         templateLoader = nil
         templatesDirectoryLocator = nil
+        templateGenerator = nil
         super.tearDown()
     }
 
@@ -42,17 +46,6 @@ final class ScaffoldCommandTests: TuistUnitTestCase {
 
     func test_overview() {
         XCTAssertEqual(ScaffoldCommand.overview, "Generates new project based on template.")
-    }
-
-    func test_fails_when_directory_not_empty() throws {
-        // Given
-        let path = FileHandler.shared.currentPath
-        try FileHandler.shared.touch(path.appending(component: "dummy"))
-
-        let result = try parser.parse([ScaffoldCommand.command, "template"])
-
-        // Then
-        XCTAssertThrowsSpecific(try subject.run(with: result), ScaffoldCommandError.nonEmptyDirectory(path))
     }
 
     func test_fails_when_template_not_found() throws {
@@ -93,5 +86,82 @@ final class ScaffoldCommandTests: TuistUnitTestCase {
         // Then
         XCTAssertThrowsError(try subject.parse(with: parser,
                                                arguments: [ScaffoldCommand.command, "template", "--name", "Test"]))
+    }
+
+    func test_fails_when_required_attribute_not_provided() throws {
+        // Given
+        templateLoader.loadTemplateStub = { _ in
+            Template.test(attributes: [.required("required")])
+        }
+
+        templatesDirectoryLocator.templateDirectoriesStub = { _ in
+            [try self.temporaryPath().appending(component: "template")]
+        }
+
+        let arguments = [ScaffoldCommand.command, "template"]
+        _ = try subject.parse(with: parser, arguments: arguments)
+
+        // When
+        let result = try parser.parse(arguments)
+
+        // Then
+        XCTAssertThrowsSpecific(try subject.run(with: result),
+                                ScaffoldCommandError.attributeNotProvided("required"))
+    }
+
+    func test_optional_attribute_is_taken_from_template() throws {
+        // Given
+        templateLoader.loadTemplateStub = { _ in
+            Template.test(attributes: [.optional("optional", default: "optionalValue")])
+        }
+
+        templatesDirectoryLocator.templateDirectoriesStub = { _ in
+            [try self.temporaryPath().appending(component: "template")]
+        }
+
+        var generateAttributes: [String: String] = [:]
+        templateGenerator.generateStub = { _, _, attributes in
+            generateAttributes = attributes
+        }
+
+        let arguments = [ScaffoldCommand.command, "template"]
+        _ = try subject.parse(with: parser, arguments: arguments)
+        let result = try parser.parse(arguments)
+
+        // When
+        try subject.run(with: result)
+
+        // Then
+        XCTAssertEqual(["optional": "optionalValue"],
+                       generateAttributes)
+    }
+
+    func test_attributes_are_passed_to_generator() throws {
+        // Given
+        templateLoader.loadTemplateStub = { _ in
+            Template.test(attributes: [.optional("optional", default: ""),
+                                       .required("required")])
+        }
+
+        templatesDirectoryLocator.templateDirectoriesStub = { _ in
+            [try self.temporaryPath().appending(component: "template")]
+        }
+
+        var generateAttributes: [String: String] = [:]
+        templateGenerator.generateStub = { _, _, attributes in
+            generateAttributes = attributes
+        }
+
+        let arguments = [ScaffoldCommand.command, "template", "--optional", "optionalValue", "--required", "requiredValue"]
+        _ = try subject.parse(with: parser, arguments: arguments)
+        let result = try parser.parse(arguments)
+
+        // When
+        try subject.run(with: result)
+
+        // Then
+        XCTAssertEqual(["optional": "optionalValue",
+                        "required": "requiredValue"],
+                       generateAttributes)
     }
 }