+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 - Add Mint up https://github.com/tuist/tuist/pull/1131 by @mollyIV
 
+### Removed
+
+- `Graphing` protocol https://github.com/tuist/tuist/pull/1128 by @pepibumur
+
 ### Changed
 
 - Optimize `TargetNode`'s set operations https://github.com/tuist/tuist/pull/1095 by @kwridan
@@ -23,6 +27,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ### Added
 
+- `scaffold` command to generate user-defined templates https://github.com/tuist/tuist/pull/1126 by @fortmarek
 - New `ProjectDescription` models for `scaffold` command https://github.com/tuist/tuist/pull/1082 by @fortmarek
 - Allow specifying Project Organization name via new `organizationName` parameter to `Project` initializer or via `Config` new GenerationOption. https://github.com/tuist/tuist/pull/1062 by @c0diq
 - `tuist lint` command https://github.com/tuist/tuist/pull/1043 by @pepibumur.